+++ src/main/java/interface_adapter/save_game/SaveGameState.java
@@ -3,10 +3,11 @@ package interface_adapter.save_game;
 import entities.Pet;
 
 public class SaveGameState {
-    private Double timeLeft = 0.0;
+    private int timeLeft = 0;
     private Pet currentPet = null;
     private String petError;
 
+    // placeholder until can get actual state from pet room view.
     public SaveGameState() {
         currentPet = new Pet(100, 100, 100, 100);
         currentPet.setPetSpritePath("");
@@ -14,7 +15,7 @@ public class SaveGameState {
         currentPet.setPetSpritePath("src/main/java/resources/dog_room_basic.jpg");
     }
 
-    public Double getTimeLeft() {
+    public int getTimeLeft() {
         return timeLeft;
     }
 
@@ -24,7 +25,7 @@ public class SaveGameState {
 
     public String getPetError() { return petError; }
 
-    public void setTimeLeft(double timeLeft) {
+    public void setTimeLeft(int timeLeft) {
         this.timeLeft = timeLeft;
     }
